vpci: use pcidevs locking to protect MMIO handlers
vPCI MMIO handlers are accessing pdevs without protecting this access
with pcidevs_{lock|unlock}. This is not a problem as of now as these
are only used by Dom0. But, towards vPCI is used also for guests, we need
to properly protect pdev and pdev->vpci from being removed while still
in use.

This patch adds ASSERTs in the code to check that the rwlock is taken
and in appropriate mode. Some of such checks require changes to the
initialization of local variables which may be accessed before the
ASSERT checks the locking. For example see init_bars and mask_write.

Also do not acquire pdev->vpci->lock spinlock for vpci_{add|remove}_register
as these are now protected by the write lock.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>