vpci: prevent BAR overlap check clash
While checking for overlapping BARs in modify_bars it is possible
that there are two callers of the later both holding the pcidevs
read lock acquired in vpci_write. This clash may produce wrong
detection of BAR overlaps. To prevent that add means to detect
such a possibility and acquire the lock in exclusive (write) mode
if detected.

1. Always start vpci_write assuming we need a read lock. At the
same time there are no means to upgrade the read lock to the write
one in case the assumption is wrong. So, if this is the case, then
we need to drop the read lock and re-start the attempt for acquiring
the write lock. It is possible that, after we drop the read lock,
pdev gets removed or re-created in between.

2. The need for the write lock comes from the fact that writing the
command register or ROM BAR register may trigger modify_bars to run,
which in turn may access multiple pdevs while checking for the
existing BAR's overlap.

All other code, which doesn't lead to pdev->vpci destruction and does not
access multiple pdevs at the same time, can still use a combination of the
read lock and pdev->vpci->lock.

3. For better write lock requirement detection optimize the code by
caching offset of the ROM BAR register in vpci->header->rom_reg which
depends on header's type.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>